:rocket: Début de la maj

+++ cockpit/modules/Cockpit/Controller/Media.php
@@ -1,4 +1,12 @@
 <?php
+/**
+ * This file is part of the Cockpit project.
+ *
+ * (c) Artur Heinze - 🅰🅶🅴🅽🆃🅴🅹🅾, http://agentejo.com
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
 
 namespace Cockpit\Controller;
 
@@ -8,7 +16,7 @@ class Media extends \Cockpit\AuthController {
 
     public function api() {
 
-        $cmd       = $this->param("cmd", false);
+        $cmd       = $this->param('cmd', false);
         $mediapath = $this->module('cockpit')->getGroupVar('finder.path', '');
 
         $this->root = rtrim($this->app->path("site:{$mediapath}"), '/');
@@ -25,7 +33,7 @@ class Media extends \Cockpit\AuthController {
 
     protected function ls() {
 
-        $data     = array("folders"=>array(), "files"=>array());
+        $data     = ['folders'=>[], 'files'=>[]];
         $toignore = [
             '.svn', '_svn', 'cvs', '_darcs', '.arch-params', '.monotone', '.bzr', '.git', '.hg',
             '.ds_store', '.thumb', '.idea'
@@ -35,10 +43,10 @@ class Media extends \Cockpit\AuthController {
         $sitefolder = $this->app->path('site:');
         $isSuperAdmin = $this->module('cockpit')->isSuperAdmin();
 
-        if ($path = $this->param("path", false)){
+        if ($path = $this->_getPathParameter()){
 
             $dir = $this->root.'/'.trim($path, '/');
-            $data["path"] = $dir;
+            $data['path'] = $dir;
 
             if (file_exists($dir)){
 
@@ -53,18 +61,20 @@ class Media extends \Cockpit\AuthController {
 
                     $isDir = $file->isDir();
 
-                    $data[$file->isDir() ? "folders":"files"][] = array(
-                        "is_file" => !$isDir,
-                        "is_dir" => $isDir,
-                        "is_writable" => is_writable($file->getPathname()),
-                        "name" => $filename,
-                        "path" => trim($path.'/'.$file->getFilename(), '/'),
-                        "rel_site_path" => trim(str_replace($sitefolder, '', $file->getPathname()), '/'),
-                        "url"  => $this->app->pathToUrl($file->getPathname()),
-                        "size" => $isDir ? "" : $this->app->helper("utils")->formatSize($file->getSize()),
-                        "ext"  => $isDir ? "" : strtolower($file->getExtension()),
-                        "lastmodified" => $file->isDir() ? "" : date("d.m.y H:i", $file->getMTime()),
-                    );
+                    $data[$file->isDir() ? 'folders':'files'][] = [
+                        'is_file' => !$isDir,
+                        'is_dir' => $isDir,
+                        'is_writable' => is_writable($file->getPathname()),
+                        'name' => $filename,
+                        'path' => trim($path.'/'.$file->getFilename(), '/'),
+                        'rel_site_path' => trim(str_replace($sitefolder, '', $file->getPathname()), '/'),
+                        'url'  => $this->app->pathToUrl($file->getPathname()),
+                        'size' => $isDir ? '' : $this->app->helper('utils')->formatSize($file->getSize()),
+                        'filesize' => $isDir ? '' : $file->getSize(),
+                        'ext'  => $isDir ? '' : strtolower($file->getExtension()),
+                        'lastmodified' => $file->isDir() ? '' : date('d.m.y H:i', $file->getMTime()),
+                        'modified' => $file->isDir() ? '' : $file->getMTime(),
+                    ];
                 }
             }
         }
@@ -74,8 +84,11 @@ class Media extends \Cockpit\AuthController {
 
     protected function upload() {
 
-        $files      = isset($_FILES['files']) ? $_FILES['files'] : [];
-        $path       = $this->param('path', false);
+        $path       = $this->_getPathParameter();
+
+        if (!$path) return false;
+
+        $files      = $_FILES['files'] ?? [];
         $targetpath = $this->root.'/'.trim($path, '/');
         $uploaded   = [];
         $failed     = [];
@@ -85,16 +98,17 @@ class Media extends \Cockpit\AuthController {
         $_failed    = [];
 
         if (isset($files['name']) && $path && file_exists($targetpath)) {
+
             for ($i = 0; $i < count($files['name']); $i++) {
 
                 // clean filename
                 $clean = preg_replace('/[^a-zA-Z0-9-_\.]/','', str_replace(' ', '-', $files['name'][$i]));
 
-                if (!$files['error'][$i] && move_uploaded_file($files['tmp_name'][$i], $targetpath.'/'.$clean)) {
+                if (!$files['error'][$i] && $this->_isFileTypeAllowed($clean) && move_uploaded_file($files['tmp_name'][$i], $targetpath.'/'.$clean)) {
                     $uploaded[]  = $files['name'][$i];
                     $_uploaded[] = $targetpath.'/'.$clean;
                 } else {
-                    $failed[]    = $files['name'][$i];
+                    $failed[]    = ['file' => $files['name'][$i], 'error' => $files['error'][$i]];
                     $_failed[]   = $targetpath.'/'.$clean;
                 }
             }
@@ -107,7 +121,10 @@ class Media extends \Cockpit\AuthController {
 
     protected function createfolder() {
 
-        $path = $this->param('path', false);
+        $path = $this->_getPathParameter();
+
+        if (!$path) return false;
+
         $name = $this->param('name', false);
         $ret  = false;
 
@@ -115,26 +132,29 @@ class Media extends \Cockpit\AuthController {
             $ret = mkdir($this->root.'/'.trim($path, '/').'/'.$name);
         }
 
-        return json_encode(array("success"=>$ret));
+        return json_encode(['success' => $ret]);
     }
 
     protected function createfile() {
 
-        $path = $this->param('path', false);
+        $path = $this->_getPathParameter();
+
+        if (!$path) return false;
+
         $name = $this->param('name', false);
         $ret  = false;
 
-        if ($name && $path) {
-            $ret = @file_put_contents($this->root.'/'.trim($path, '/').'/'.$name, "");
+        if ($name && $this->_isFileTypeAllowed($name) && $path) {
+            $ret = @file_put_contents($this->root.'/'.trim($path, '/').'/'.$name, '');
         }
 
-        return json_encode(array("success"=>$ret));
+        return json_encode(['success' => $ret]);
     }
 
 
     protected function removefiles() {
 
-        $paths     = (array)$this->param('paths', array());
+        $paths     = (array)$this->param('paths', []);
         $deletions = [];
 
         foreach ($paths as $path) {
@@ -154,7 +174,7 @@ class Media extends \Cockpit\AuthController {
 
         $this->app->trigger('cockpit.media.removefiles', [$deletions]);
 
-        return json_encode(array("success"=>true));
+        return json_encode(["success"=>true]);
     }
 
     protected function _rrmdir($dir) {
@@ -174,10 +194,13 @@ class Media extends \Cockpit\AuthController {
 
     protected function rename() {
 
-        $path = $this->param('path', false);
+        $path = $this->_getPathParameter();
+
+        if (!$path) return false;
+
         $name = $this->param('name', false);
 
-        if ($path && $name) {
+        if ($path && $name && $this->_isFileTypeAllowed($name)) {
             $source = $this->root.'/'.trim($path, '/');
             $target = dirname($source).'/'.$name;
 
@@ -185,12 +208,15 @@ class Media extends \Cockpit\AuthController {
             $this->app->trigger('cockpit.media.rename', [$source, $target]);
         }
 
-        return json_encode(array("success"=>true));
+        return json_encode(["success"=>true]);
     }
 
     protected function readfile() {
 
-        $path = $this->param('path', false);
+        $path = $this->_getPathParameter();
+
+        if (!$path) return false;
+
         $file = $this->root.'/'.trim($path, '/');
 
         if ($path && file_exists($file)) {
@@ -202,7 +228,10 @@ class Media extends \Cockpit\AuthController {
 
     protected function writefile() {
 
-        $path    = $this->param('path', false);
+        $path    = $this->_getPathParameter();
+
+        if (!$path) return false;
+
         $content = $this->param('content', false);
         $file    = $this->root.'/'.trim($path, '/');
         $ret     = false;
@@ -211,14 +240,16 @@ class Media extends \Cockpit\AuthController {
             $ret = file_put_contents($file, $content);
         }
 
-        return json_encode(array("success"=>$ret));
+        return json_encode(['success' => $ret]);
     }
 
     protected function unzip() {
 
-        $return  = ['success' => false];
+        $path    = $this->_getPathParameter();
 
-        $path    = $this->param('path', false);
+        if (!$path) return false;
+
+        $return  = ['success' => false];
         $zip     = $this->param('zip', false);
 
         if ($path && $zip) {
@@ -243,24 +274,72 @@ class Media extends \Cockpit\AuthController {
 
     protected function download() {
 
-        $path = $this->param('path', false);
+        $path = $this->_getPathParameter();
+
+        if (!$path) return false;
+
         $file = $this->root.'/'.trim($path, '/');
 
         if (!$path && !file_exists($file)) {
             $this->app->stop();
         }
 
+        if (is_dir($file)) {
+            return $this->downloadfolder();
+        }
+
         $pathinfo = $path_parts = pathinfo($file);
 
-        header("Pragma: public");
-        header("Expires: 0");
-        header("Cache-Control: must-revalidate, post-check=0, pre-check=0");
-        header("Cache-Control: private",false);
-        header("Content-Type: application/force-download");
-        header("Content-Disposition: attachment; filename=\"".$pathinfo["basename"]."\";" );
-        header("Content-Transfer-Encoding: binary");
-        header("Content-Length: ".filesize($file));
-        readfile($file);
+        header('Pragma: public');
+        header('Expires: 0');
+        header('Cache-Control: must-revalidate, post-check=0, pre-check=0');
+        header('Cache-Control: private', false);
+        header('Content-Type: application/force-download');
+        header('Content-Disposition: attachment; filename="'.$pathinfo["basename"].'";' );
+        header('Content-Transfer-Encoding: binary');
+        header('Content-Length: '.filesize($file));
+
+        //readfile($file);
+
+        $handle = fopen($file, 'rb');
+
+        while (!feof($handle)) {
+            echo fread($handle, 1000);
+        }
+
+        fclose($handle);
+
+        $this->app->stop();
+    }
+
+    protected function downloadfolder() {
+
+        $path   = $this->_getPathParameter();
+
+        if (!$path) return false;
+
+        $folder = $this->root.'/'.trim($path, '/');
+
+        if (!$path && !file_exists($folder)) {
+            $this->app->stop();
+        }
+
+        header('X-Accel-Buffering: no');
+
+        $prefix = basename($path);
+        $files  = new \RecursiveIteratorIterator(new \RecursiveDirectoryIterator($folder), \RecursiveIteratorIterator::LEAVES_ONLY);
+        $zip    = new \ZipStream\ZipStream("{$prefix}.zip");
+
+        foreach ($files as $name => $file) {
+
+            if ($file->isDir()) continue;
+
+            $filePath = $file->getRealPath();
+            $relativePath = substr($filePath, strlen($folder) + 1);
+            $zip->addFileFromPath("{$prefix}/{$relativePath}", $filePath);
+        }
+
+        $zip->finish();
 
         $this->app->stop();
     }
@@ -301,7 +380,7 @@ class Media extends \Cockpit\AuthController {
     public function savebookmarks() {
 
         if ($bookmarks = $this->param('bookmarks', false)) {
-            $this->memory->set("mediamanager.bookmarks.".$this->user["_id"], $bookmarks);
+            $this->memory->set('mediamanager.bookmarks.'.$this->user['_id'], $bookmarks);
         }
 
         return json_encode($bookmarks);
@@ -309,7 +388,36 @@ class Media extends \Cockpit\AuthController {
 
     public function loadbookmarks() {
 
-        return json_encode($this->app->memory->get("mediamanager.bookmarks.".$this->user["_id"], ["folders"=>[], "files"=>[]]));
+        return json_encode($this->app->memory->get('mediamanager.bookmarks.'.$this->user['_id'], ['folders'=>[], 'files'=>[]]));
+    }
+
+    protected function _getPathParameter() {
+
+        $path = $this->param('path', false);
+
+        if ($path) {
+
+            $path = trim($path);
+
+            if (strpos($path, '../') !== false) {
+                $path = false;
+            }
+        }
+
+        return $path;
+    }
+
+    protected function _isFileTypeAllowed($file) {
+
+        $allowed = trim($this->module('cockpit')->getGroupVar('finder.allowed_uploads', $this->app->retrieve('allowed_uploads', '*')));
+
+        if ($allowed == '*') {
+            return true;
+        }
+
+        $allowed = str_replace([' ', ','], ['', '|'], preg_quote(is_array($allowed) ? implode(',', $allowed) : $allowed));
+
+        return preg_match("/\.({$allowed})$/i", $file);
     }
 
 }
